feat: Expose libggml in internal APIs (#1761)
* Expose libggml and refactor ctypes extension

* Only expose libggml

* Use ctypes_extensions module for libllama and libllava
